+++ ruff-lint-strict.toml
@@ -23,10 +23,10 @@ exclude = [
     "alphadia/calibration/*.py",
     "alphadia/consensus/*.py",
     "alphadia/constants/*.py",
-    "alphadia/data/alpharaw_wrapper.py",
-    "alphadia/data/bruker.py",
-    "alphadia/data/utils.py",
-    "alphadia/data/jitclasses/*.py",
+    "alphadia/raw_data/alpharaw_wrapper.py",
+    "alphadia/raw_data/bruker.py",
+    "alphadia/raw_data/utils.py",
+    "alphadia/raw_data/jitclasses/*.py",
     "alphadia/_fdrx/stats.py",
     "alphadia/_fdrx/plotting.py",
     "alphadia/_fdrx/base.py",
